ADDED Enquiry and Query delete methods

+++ src/main/java/com/MasterBranch/repository/QueryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.MasterBranch.model.Query;
 
 @Transactional
-public interface QueryRepository  extends JpaRepository<Query, Long> {
+public interface QueryRepository  extends JpaRepository<Query, Integer> {
 	List<Query> findAll();
 }
